+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+## Unreleased
+
+- add `Combo.Proxy`